mqtt-client/requirements.txt
paho.mqtt
redis
dataset
pymongo

